requirements.txt
# torch>=2.0.0
# torchvision>=0.15.0
# diffusers>=0.21.0
# transformers>=4.30.0
accelerate>=0.20.0
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.0.0
python-multipart>=0.0.6
pyyaml>=6.0
safetensors>=0.3.0
pillow>=10.0.0